testing/requirements.txt
click==8.1.3
halo==0.0.31
minio==7.1.13
pika==1.3.1
Pillow==9.4.0
pymongo==4.3.3
tqdm==4.64.1
